jewel_app/pyproject.toml
[tool.poetry]
name = "jewel-app"
version = "0.1.0"
description = ""
authors = ["joonhwan <joonhwan.lee@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
fastapi = "^0.79.0"
sqlmodel = "^0.0.6"
uvicorn = "^0.18.2"
alembic = "^1.8.1"
pydantic = {extras = ["email"], version = "^1.9.1"}
bcrypt = "^3.2.2"
passlib = "^1.7.4"
PyJWT = "^2.4.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
